tensquare/backend/requirements.txt
amqp==2.5.0
backports.csv==1.0.7
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.11
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-rest-framework-mongoengine==3.3.1
django-reversion==3.0.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
drf-haystack==1.8.5
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
kombu==4.6.3
mongoengine==0.18.2
mutagen==1.42.0
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
PyJWT==1.7.1
pymongo==3.8.0
PyMySQL==0.9.3
pysnowflake==0.1.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
qiniu==7.2.4
redis==3.2.1
requests==2.22.0
six==1.12.0
tablib==0.13.0
tornado==6.0.2
urllib3==1.25.3
vine==1.3.0
xadmin==0.6.1
xlrd==1.2.0
xlwt==1.3.0
